% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_manuscript_readme}
\alias{add_manuscript_readme}
\title{Add manuscript folder README file to project}
\usage{
add_manuscript_readme(path)
}
\arguments{
\item{path}{Character, path to project repository.}
}
\description{
Add manuscript folder README file to project
}
\keyword{internal}
