% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_materials_readme}
\alias{add_materials_readme}
\title{Add materials folder README file to project}
\usage{
add_materials_readme(path)
}
\arguments{
\item{path}{Character, path to project repository.}
}
\description{
Add materials folder README file to project
}
\keyword{internal}
