% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_preregistration_readme}
\alias{add_preregistration_readme}
\title{Add preregistration file to project}
\usage{
add_preregistration_readme(path, template)
}
\arguments{
\item{path}{Character, path to project repository.}

\item{template}{Character, preregistration template to be used.}
}
\description{
Add preregistration file to project
}
\keyword{internal}
