% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_project_readme}
\alias{add_project_readme}
\title{Add top-level project README file to project}
\usage{
add_project_readme(path)
}
\arguments{
\item{path}{Character, path to project repository.}
}
\description{
Add top-level project README file to project
}
\keyword{internal}
