% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_readme}
\alias{add_readme}
\title{Add README file to project}
\usage{
add_readme(path)
}
\arguments{
\item{path}{Character, path to project repository.}
}
\description{
Add README file to project
}
\keyword{internal}
