% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_registered_report}
\alias{add_registered_report}
\title{Add Registered Report file to project}
\usage{
add_registered_report(path)
}
\arguments{
\item{path}{Character, path to project repository.}
}
\description{
Add Registered Report file to project
}
\keyword{internal}
