% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_rproj}
\alias{add_rproj}
\title{Add .Rproj file to project}
\usage{
add_rproj(path)
}
\arguments{
\item{path}{Character, path to project repository.}
}
\description{
Add .Rproj file to project
}
\keyword{internal}
