% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_scripts_readme}
\alias{add_scripts_readme}
\title{Add scripts folder README file to project}
\usage{
add_scripts_readme(path)
}
\arguments{
\item{path}{Character, path to project repository.}
}
\description{
Add scripts folder README file to project
}
\keyword{internal}
