% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_supplement_readme}
\alias{add_supplement_readme}
\title{Add supplemental materials file}
\usage{
add_supplement_readme(path)
}
\arguments{
\item{path}{Character, path to project repository.}
}
\description{
Add supplemental materials file
}
\keyword{internal}
