% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-functions.R
\name{check_data_access}
\alias{check_data_access}
\title{Check whether data was accessed previously}
\usage{
check_data_access(data_hash)
}
\arguments{
\item{data_hash}{MD5 hash of the current data object (after any modifications).}
}
\description{
Check whether data was accessed previously
}
\keyword{internal}
