% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{copy_resource}
\alias{copy_resource}
\title{Add package resource to project}
\usage{
copy_resource(file, from, to)
}
\arguments{
\item{file}{Character, file to be copied.}

\item{from}{Character, Internal path to the file.}

\item{to}{Character, Location that the file should be copied to.}
}
\description{
Add package resource to project
}
\keyword{internal}
