% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{count_matching_tags}
\alias{count_matching_tags}
\title{Count number of existing tags on GitHub that match the currently supplied tag.}
\usage{
count_matching_tags(tag)
}
\arguments{
\item{tag}{Character, the tag that should be given to the milestone commit.}
}
\description{
Count number of existing tags on GitHub that match the currently supplied tag.
}
\keyword{internal}
