% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_project.R
\name{create_github_repo}
\alias{create_github_repo}
\title{Create a remote project repository on GitHub}
\usage{
create_github_repo(private_repo = TRUE)
}
\arguments{
\item{private_repo}{Logical, should the GitHub repository be private or
public?
Default is TRUE}
}
\value{
Nothing, this function is called for it's side-effects.
}
\description{
Creates a remote repository on GitHub if you had not already done so when
initiating the project. It automatically creates the repository for you
and pushes all the changes and milestones that you have logged so far.
}
