% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{generate_milestone_overview}
\alias{generate_milestone_overview}
\title{Generate an overview of all project milestones}
\usage{
generate_milestone_overview(copy = FALSE)
}
\arguments{
\item{copy}{Boolean. If TRUE, the bulleted overview of milestones is
copied to the clipboard.}
}
\value{
A bulleted overview of the project milestones, including the type of
milestone (e.g., first-time data access, timestamped preregistration), a link
to the GitHub commit, and (in the case of first-time data access) the unique
MD5 hash of the data read into the R environment, as well as a link to the
code that was used to read the data.
}
\description{
If you used `projectlog::log_milestones` throughout our project, this
function can be used to summarise them in a bulleted list. By default,
this function is included in the top-level README file as generated by
`projectlog`. If you did not use `projectlog` to generate your project,
you can run this function to generate the milestone overview and paste it
in a location of your choice.
}
