% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_project.R
\name{initiate_project}
\alias{initiate_project}
\title{Initiate your project folder}
\usage{
initiate_project(
  path,
  project = "single_study",
  preregistration = "empty",
  dependencies = "groundhog",
  private_repo = TRUE,
  ...
)
}
\arguments{
\item{path}{Character, the directory in which the project is initiated.}

\item{project}{Character, the type of project that will be initiated.
Options include 'single_study' for a project consisting of a single study,
'multistudy' for a project consisting of multiple studies, and
'registered_report' for a project containing a Registered Report.
Default: 'single_study'.}

\item{dependencies}{Character, indicates whether and how you want to keep track
of package dependencies. Options include \code{\link[renv]{init}},
\code{\link[groundhog]{groundhog}}, or 'none'.}

\item{private_repo}{Logical, Should the created GitHub repository be private or public?}

\item{...}{Additional arguments passed to and from functions.}
}
\value{
The initial set-up of your project folder.
}
\description{
This function creates your project folder.
}
