% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{A function to add lines to an existing quarto log.}
\usage{
logger()
}
\description{
A little package to write to my work notes Quarto. The LOG_PATH needs to point to a folder of qmds.
}
