% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{show_changes}
\alias{show_changes}
\title{Show untracked project changes}
\usage{
show_changes()
}
\value{
A character vector of file names.
}
\description{
This function is a wrapper around 'gert::git_status())'.
It shows all the files that have changed since the last commit to GitHub.
It is recommended to routinely commit untracked changes to GitHub using
the 'log_changes()' or 'log_milestone()' functions.
}
