% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-functions.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle variables in data.}
\usage{
shuffle(data, shuffle_vars, long_format, seed = seed)
}
\arguments{
\item{data}{Tibble or data.frame.}

\item{shuffle_vars}{Vector, Names of variables that should be shuffled.}

\item{long_format}{Logical, Is data in long-format (TRUE) or in wide-format (FALSE)?}

\item{seed}{Character, Random seed to make sure shuffling is identical in subsequent `read_data()` calls.}
}
\description{
Shuffle variables in data.
}
\keyword{internal}
