% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{validate_commit}
\alias{validate_commit}
\title{Check whether milestone can be committed to GitHub.}
\usage{
validate_commit(commit_message)
}
\arguments{
\item{commit_message}{Character, message for the milestone commit.}
}
\description{
Check whether milestone can be committed to GitHub.
}
\keyword{internal}
