% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-functions.R
\name{validate_fun}
\alias{validate_fun}
\title{Validate read function used in projectlog::initiate_project()}
\usage{
validate_fun(file, fun)
}
\arguments{
\item{file}{Character, path to data file.}

\item{fun}{Character, name of function to read data.}
}
\description{
Validate read function used in projectlog::initiate_project()
}
\keyword{internal}
