% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{validate_tag}
\alias{validate_tag}
\title{Check whether supplied tag is valid}
\usage{
validate_tag(tag)
}
\arguments{
\item{tag}{Character, a tag to be used in the milestone commit.}
}
\description{
Check whether supplied tag is valid
}
\keyword{internal}
