% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-files.R
\name{add_project_readme}
\alias{add_project_readme}
\title{Add top-level project README file to project}
\usage{
add_project_readme(path = getwd())
}
\arguments{
\item{path}{Character, path to project repository.}
}
\value{
A README.Rmd file at the highest level of your repository
}
\description{
If you did not initiate your project using `projectlog::initiate_project()`, you can use this function to generate a README file that will be placed at the highest level of your repository.
It will include example text for a description of your project, an overview of the (most important) folders, and code to generate an overview of your project milestones.
}
