% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_project.R
\name{initiate_project}
\alias{initiate_project}
\title{Initiate your project folder}
\usage{
initiate_project(
  path,
  project = "single_study",
  preregistration = "empty",
  dependencies = "groundhog",
  create_github = "yes",
  private_repo = "yes",
  ...
)
}
\arguments{
\item{path}{Character, the directory in which the project is initiated.}

\item{project}{Character, the type of project that will be initiated.
Options include 'single_study' for a project consisting of a single study,
'multistudy' for a project consisting of multiple studies, and
'registered_report' for a project containing a Registered Report.
Default: 'single_study'.}

\item{preregistration}{Character, The preregistration template that you want to use.}

\item{dependencies}{Character, indicates whether and how you want to keep track
of package dependencies. Options include "renv",
"groundhog", or "none".
Default: 'groundhog'}

\item{create_github}{Character, should a remote repository be created on GitHub?
If 'yes', a remote repository is created with the same name as the local project,
and all current changes are pushed to this remote repository. If 'no', git is
only set up locally.
Default: 'yes'}

\item{private_repo}{Logical, Should the created GitHub repository be private or public?
Default: 'yes'}

\item{...}{Additional arguments passed to and from functions.}
}
\value{
The initial set-up of your project repository.
}
\description{
This function creates your project folder.
}
