% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{validate_files}
\alias{validate_files}
\title{Check whether files to be logged are valid.}
\usage{
validate_files(files)
}
\arguments{
\item{files}{Vector of files to be logged.}
}
\description{
Check whether files to be logged are valid.
}
\keyword{internal}
